+++ R/shiny-server-params.R
@@ -1,6 +1,9 @@
-retrieve_params <- function(x) {
+# © 2019 JAMSTEC
+retrieve_params <- function(x, xmap) {
   shiny::req(x)
   y <- attr(x, "params", exact = TRUE)
-  y$beta <- y$beta / 100
+  y$beta <- y$beta *
+    attr(xmap, "dwell") * 1e-3 *
+    attr(xmap, "current") * 1e+6
   y
 }